[REFACTOR] game.js
